Finally fixed the time zone problems in the admin interface.
For future reference, all times are stored in the database as UTC.
Although django will usually handle the conversion automatically,
it didn't do so in this case because I was displaying the datetimes
manually.

Also some stuff on testing and user permissions. This needs some
major revision. I suspect authenticate is not working properly in
tastypie.